src/main/java/com/kodilla/library/controller/UserController.java
package com.kodilla.library.controller;

import com.kodilla.library.domain.dto.UserCreationDto;
import com.kodilla.library.domain.dto.UserDto;
import com.kodilla.library.mapper.UserMapper;
import com.kodilla.library.service.UserService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.*;
import com.kodilla.library.controller.controllerExceptions.LibraryDatabaseException;
import java.util.List;

@CrossOrigin(origins = "*")
@RestController
@RequestMapping("/library")
public class UserController {
    @Autowired
    private UserMapper userMapper;

    @Autowired
    private UserService userService;

    @RequestMapping(method = RequestMethod.GET, value = "/users")
    public List<UserDto> getUsers(){
        return userMapper.mapToUserDtoList(userService.getAllUsers());
    }

    @RequestMapping(method = RequestMethod.GET, value = "/users/{id}")
    public UserDto getUser(@PathVariable Long id) throws LibraryDatabaseException {
        return userMapper.mapToUserDto(userService.findUserById(id));
    }

    @RequestMapping(method = RequestMethod.POST, value = "/users")
    public void addUser(@RequestBody UserCreationDto userCreationDto){
        userService.addUser(userMapper.mapToUserFromCreationDto(userCreationDto));
    }

    @RequestMapping(method = RequestMethod.DELETE, value = "/users/{id}")
    public void deleteUser(@PathVariable Long id){
        userService.deleteUserById(id);
    }
}
